Refactor
Remove unused code
In DomainDao change totalCount return type from long to int

+++ database/src/main/java/ir/comprehensive/database/provider/service/AbstractDomainDao.java
@@ -101,7 +101,7 @@ public abstract class AbstractDomainDao<M extends DomainModel<I>, R extends Reco
     }
 
     @Override
-    public long totalCount() throws SearchingException {
+    public int totalCount() throws SearchingException {
         return context.fetchCount(this.table);
     }
